+++ changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 1.0.6
+
 - Added discount for PRO upgrade purchase if you already have PREMIUM
 - Add admin page for button type
 - Adding multiple pages in inventory, currently you can go up to 10 pages